refactore the whole scripts adding validation on installation

+++ scripts/code.sh
@@ -1,11 +1,34 @@
 #!/bin/bash
 
-echo "<> Installing VS Code <>"
+extension_installation() {
+  echo "should the extensions be installed? (y/n)"
+  read install_extension
+  if echo "$install_extension" | grep -iq "^y"; then
+    code --install-extension Shan.code-settings-sync
+  fi
+}
 
-wget -q https://packages.microsoft.com/keys/microsoft.asc -O- | sudo apt-key add -
+code_installation() {
+  if [ $(which code | grep -c "code") -eq 0 ]; then
+    wget -q https://packages.microsoft.com/keys/microsoft.asc -O- | sudo apt-key add -
+    sudo add-apt-repository "deb [arch=amd64] https://packages.microsoft.com/repos/vscode stable main"
+    sudo apt install -y code
+  else
+    echo "vs code already installed!"
+  fi
+}
 
-sudo add-apt-repository "deb [arch=amd64] https://packages.microsoft.com/repos/vscode stable main"
+echo "<> Do you want to install VS code? (y/n)"
+read install_code
 
-sudo apt install -y code
+if echo "$install_code" | grep -iq "^y"; then
+  echo "<> Installing VS Code <>"
 
-code --install-extension Shan.code-settings-sync
+  code_installation
+
+  extension_installation
+
+  echo "\n\n"
+else
+  echo "VS Code will not be installed XD \n\n"
+fi
