+++ scripts/slack.sh
@@ -1,8 +1,22 @@
 #!/bin/bash
-echo "<> Download and Installing Slack <>"
 
-wget https://downloads.slack-edge.com/linux_releases/slack-desktop-4.4.2-amd64.deb -P "$PWD/.files"
+slack_installation() {
+  if [ $(which slack | grep -c "slack") -eq 0 ]; then
+    echo "<> Download and Installing Slack <>"
+    wget https://downloads.slack-edge.com/linux_releases/slack-desktop-4.4.2-amd64.deb -P "$PWD/.files"
+    sudo dpkg -i "$PWD/.files/slack-desktop-4.4.2-amd64.deb"
+    sudo rm -rf $PWD/.files/slack-desktop-4.4.2-amd64.deb
+  else
+    echo "Slack already installed!"
+  fi
+}
 
-sudo dpkg -i "$PWD/.files/slack-desktop-4.4.2-amd64.deb"
+echo "<> Do you want to install Slack? (y/n)"
+read install_slack
 
-sudo rm -rf $PWD/.files/slack-desktop-4.4.2-amd64.deb
+if echo "$install_slack" | grep -iq "^y"; then
+  slack_installation
+  echo "\n\n"
+else
+  echo "Slack will not be installed XD \n\n"
+fi
